AutomatedScannerTest/requirements.txt
beets==2.2.0
colorama==0.4.6
confuse==2.0.1
filetype==1.2.0
jellyfish==1.1.3
mediafile==0.13.0
munkres==1.1.4
musicbrainzngs==0.7.1
mutagen==1.47.0
pillow==11.1.0
pip==25.0.1
platformdirs==4.3.6
pytz==2025.1
PyVISA==1.14.1
PyVISA-py==0.7.2
PyYAML==6.0.2
typing_extensions==4.12.2
Unidecode==1.3.8
